Skeleton for 3rd iteration of project
