+++ src/main/java/com/spartaglobal/RabbitSimulator/Breeder/Breeder.java
@@ -1,0 +1,9 @@
+package com.spartaglobal.RabbitSimulator.Breeder;
+
+public interface Breeder {
+
+    public static void makeNewAnimals() {
+
+    }
+
+}
