+++ src/main/java/com/spartaglobal/RabbitSimulator/Breeder/FoxBreeder.java
@@ -1,0 +1,7 @@
+package com.spartaglobal.RabbitSimulator.Breeder;
+
+public class FoxBreeder implements Breeder {
+
+
+
+}
