+++ src/main/java/com/spartaglobal/RabbitSimulator/Population/FoxPopulation.java
@@ -1,0 +1,6 @@
+package com.spartaglobal.RabbitSimulator.Population;
+
+public class FoxPopulation extends Population {
+
+
+}
