+++ src/main/java/com/spartaglobal/RabbitSimulator/Predator/Eater.java
@@ -1,0 +1,9 @@
+package com.spartaglobal.RabbitSimulator.Predator;
+
+public class Eater {
+
+    public static void foxesEatRabbits() {
+
+    }
+
+}
