--- src/main/resources/monthlyReport.txt
@@ -1,181 +1,0 @@
--------------------------------
-Monthly Report for Month 1
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 2
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 3
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 4
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 5
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 6
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 7
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 8
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 9
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 10
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 11
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 12
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 13
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 14
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 15
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 16
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 17
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 18
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 19
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 20
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 21
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 22
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 23
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 24
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 25
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 26
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 27
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 28
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 29
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
---------------------------------------------------------------
-Monthly Report for Month 30
-Total population of rabbits: null
-Population of females: null
-Population of males: null
-Population of total expired rabbits: null
--------------------------------
